requirements.txt
autopep8==1.4.4
pytest==5.1.2
pytest-cov==2.7.1
tox==3.14.0
python-Levenshtein==0.12.0
mypy==0.720

# Not the latest for py2
astroid==2.2.5
isort==4.3.21
pylint==2.3.1
